src/compute-plane-services/byoo-otel-collector/generator/pyproject.toml
# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
# SPDX-License-Identifier: Apache-2.0

[project]
name = "config-generator"
version = "0.1.0"
description = "Config Template and docs generator for byoo-otel-collector"
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.0.3",
    "PyYAML>=5.4.1",
]
